Expose Parsers and their Registries of Deep Links
To be able to access all deeplinks captured by DLD,
without needing reflection, add getRegistry() to Parser and
change Parser to an Abstract Class.

Update tests to match.

To be able to access a destination, add ComponentName to
DeepLinkResult.

Potentially, storing more information, like Extras could
allow for unit tests to assert that URI parameters have
been parsed as extras. To enable passing Android-framework
things in DeepLinkResult, @BenSchwab pointed out that it
may be possible to move DeepLinkResult to an Android
module (out of deeplinkdispatch-base).